refactor(service): decompose context.rs into focused submodules (KAN-766)
